F(floats) for smooth motion

+++ src/player.cpp
@@ -14,7 +14,7 @@ Player::Player() : Entity() {
     _sprite.AddAnimation("run" , 192, 100, 16, 28, 4);
 }
 
-Player::Player(int center_x, int center_y) : Player() {
+Player::Player(float center_x, float center_y) : Player() {
     _body   = {
         center_x - 48 / 2,
         center_y - 84 / 2,
